clean up and fix finalnode semantics
