+++ test/testfiles/decision_node_test.nt
@@ -83,8 +83,8 @@
 <https://bbn.com/scratch/execute_0> <http://sbols.org/v3#hasNamespace> <https://bbn.com/scratch/> .
 <https://bbn.com/scratch/execute_0> <http://www.w3.org/1999/02/22-rdf-syntax-ns#type> <http://bioprotocols.org/paml#BehaviorExecution> .
 <https://bbn.com/scratch/execute_0> <http://www.w3.org/1999/02/22-rdf-syntax-ns#type> <http://sbols.org/v3#TopLevel> .
-<https://bbn.com/scratch/execute_0> <http://www.w3.org/ns/prov#endedAtTime> "2022-06-15T13:48:44.050848"^^<http://www.w3.org/2001/XMLSchema#dateTime> .
-<https://bbn.com/scratch/execute_0> <http://www.w3.org/ns/prov#startedAtTime> "2022-06-15T13:48:44.050839"^^<http://www.w3.org/2001/XMLSchema#dateTime> .
+<https://bbn.com/scratch/execute_0> <http://www.w3.org/ns/prov#endedAtTime> "2022-06-15T21:10:20.202819"^^<http://www.w3.org/2001/XMLSchema#dateTime> .
+<https://bbn.com/scratch/execute_0> <http://www.w3.org/ns/prov#startedAtTime> "2022-06-15T21:10:20.202816"^^<http://www.w3.org/2001/XMLSchema#dateTime> .
 <https://bbn.com/scratch/pHMeterCalibrated/OrderedPropertyValue1/Parameter1/LiteralInteger1> <http://bioprotocols.org/uml#integerValue> "1"^^<http://www.w3.org/2001/XMLSchema#integer> .
 <https://bbn.com/scratch/pHMeterCalibrated/OrderedPropertyValue1/Parameter1/LiteralInteger1> <http://sbols.org/v3#displayId> "LiteralInteger1" .
 <https://bbn.com/scratch/pHMeterCalibrated/OrderedPropertyValue1/Parameter1/LiteralInteger1> <http://www.w3.org/1999/02/22-rdf-syntax-ns#type> <http://bioprotocols.org/uml#LiteralInteger> .
@@ -164,6 +164,11 @@
 <https://bbn.com/scratch/test_execution/ActivityNodeExecution2> <http://sbols.org/v3#displayId> "ActivityNodeExecution2" .
 <https://bbn.com/scratch/test_execution/ActivityNodeExecution2> <http://www.w3.org/1999/02/22-rdf-syntax-ns#type> <http://bioprotocols.org/paml#ActivityNodeExecution> .
 <https://bbn.com/scratch/test_execution/ActivityNodeExecution2> <http://www.w3.org/1999/02/22-rdf-syntax-ns#type> <http://sbols.org/v3#Identified> .
+<https://bbn.com/scratch/test_execution/ActivityNodeExecution3> <http://bioprotocols.org/paml#incomingFlow> <https://bbn.com/scratch/test_execution/ActivityEdgeFlow4> .
+<https://bbn.com/scratch/test_execution/ActivityNodeExecution3> <http://bioprotocols.org/paml#node> <https://bbn.com/scratch/decision_node_test/FlowFinalNode1> .
+<https://bbn.com/scratch/test_execution/ActivityNodeExecution3> <http://sbols.org/v3#displayId> "ActivityNodeExecution3" .
+<https://bbn.com/scratch/test_execution/ActivityNodeExecution3> <http://www.w3.org/1999/02/22-rdf-syntax-ns#type> <http://bioprotocols.org/paml#ActivityNodeExecution> .
+<https://bbn.com/scratch/test_execution/ActivityNodeExecution3> <http://www.w3.org/1999/02/22-rdf-syntax-ns#type> <http://sbols.org/v3#Identified> .
 <https://bbn.com/scratch/test_execution/Association1> <http://sbols.org/v3#displayId> "Association1" .
 <https://bbn.com/scratch/test_execution/Association1> <http://www.w3.org/1999/02/22-rdf-syntax-ns#type> <http://sbols.org/v3#Identified> .
 <https://bbn.com/scratch/test_execution/Association1> <http://www.w3.org/1999/02/22-rdf-syntax-ns#type> <http://www.w3.org/ns/prov#Association> .
@@ -178,6 +183,7 @@
 <https://bbn.com/scratch/test_execution> <http://bioprotocols.org/paml#completedNormally> "true"^^<http://www.w3.org/2001/XMLSchema#boolean> .
 <https://bbn.com/scratch/test_execution> <http://bioprotocols.org/paml#execution> <https://bbn.com/scratch/test_execution/ActivityNodeExecution1> .
 <https://bbn.com/scratch/test_execution> <http://bioprotocols.org/paml#execution> <https://bbn.com/scratch/test_execution/ActivityNodeExecution2> .
+<https://bbn.com/scratch/test_execution> <http://bioprotocols.org/paml#execution> <https://bbn.com/scratch/test_execution/ActivityNodeExecution3> .
 <https://bbn.com/scratch/test_execution> <http://bioprotocols.org/paml#execution> <https://bbn.com/scratch/test_execution/CallBehaviorExecution1> .
 <https://bbn.com/scratch/test_execution> <http://bioprotocols.org/paml#flow> <https://bbn.com/scratch/test_execution/ActivityEdgeFlow1> .
 <https://bbn.com/scratch/test_execution> <http://bioprotocols.org/paml#flow> <https://bbn.com/scratch/test_execution/ActivityEdgeFlow2> .
@@ -188,6 +194,6 @@
 <https://bbn.com/scratch/test_execution> <http://sbols.org/v3#hasNamespace> <https://bbn.com/scratch/> .
 <https://bbn.com/scratch/test_execution> <http://www.w3.org/1999/02/22-rdf-syntax-ns#type> <http://bioprotocols.org/paml#ProtocolExecution> .
 <https://bbn.com/scratch/test_execution> <http://www.w3.org/1999/02/22-rdf-syntax-ns#type> <http://sbols.org/v3#TopLevel> .
-<https://bbn.com/scratch/test_execution> <http://www.w3.org/ns/prov#endedAtTime> "2022-06-15T13:49:26.474080"^^<http://www.w3.org/2001/XMLSchema#dateTime> .
+<https://bbn.com/scratch/test_execution> <http://www.w3.org/ns/prov#endedAtTime> "2022-06-15T21:10:20.221458"^^<http://www.w3.org/2001/XMLSchema#dateTime> .
 <https://bbn.com/scratch/test_execution> <http://www.w3.org/ns/prov#qualifiedAssociation> <https://bbn.com/scratch/test_execution/Association1> .
-<https://bbn.com/scratch/test_execution> <http://www.w3.org/ns/prov#startedAtTime> "2022-06-15T13:48:44.045269"^^<http://www.w3.org/2001/XMLSchema#dateTime> .
+<https://bbn.com/scratch/test_execution> <http://www.w3.org/ns/prov#startedAtTime> "2022-06-15T21:10:20.197121"^^<http://www.w3.org/2001/XMLSchema#dateTime> .
